chore: Clean Bean and relocate enums - `CartBean`: Remove `product_id` - `OrderBean`: Remove unused `user_id` - `PaymentBean`: Remove unused `order_id`, change ManyToOne to OneToOne - `ProductBean`: Remove `promotion_id` and `category_id` - Move Enums to dedicated `enums` package
