+++ src/main/java/com/example/demo/bean/CartBean.java
@@ -12,12 +12,6 @@ public class CartBean {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     
-    @Column(name = "user_id", nullable = false)
-    private Long userId;
-    
-    @Column(name = "product_id", nullable = false)
-    private Long productId;
-    
     @Column(nullable = false)
     private Integer quantity;
     
@@ -35,7 +29,4 @@ public class CartBean {
     @JoinColumn(name = "user_id", insertable = false, updatable = false)
     private UserBean user;
     
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "product_id", insertable = false, updatable = false)
-    private ProductBean product;
 }
