+++ src/main/java/com/example/demo/bean/CategoryBean.java
@@ -15,4 +15,6 @@ public class CategoryBean {
     private String name;
     
     private String description;
+
+    
 }
