+++ src/main/java/com/example/demo/bean/TagBean.java
@@ -14,3 +14,4 @@ public class TagBean {
     @Column(nullable = false)
     private String name;
 }
+
