+++ src/main/java/com/example/demo/enums/PaymentStatus.java
@@ -1,4 +1,4 @@
-package com.example.demo.bean;
+package com.example.demo.enums;
 
 public enum PaymentStatus {
     PENDING,
